+++ libvast/src/command.cpp
@@ -350,35 +350,6 @@ caf::error parse_impl(invocation& result, const command& cmd,
   return parse_impl(result, **i, position + 1, last, target);
 }
 
-void fixup_options(invocation& inv) {
-  using namespace std::string_literals;
-  auto move_option_arg = [&](const char* cmd, const char* format,
-                             const char* opt) {
-    auto path = "vast."s + cmd + '.' + format + '.' + opt;
-    if (auto x = caf::get_if<std::string>(&inv.options, path)) {
-      // The logger isn't initialized yet.
-#if VAST_LOG_LEVEL >= VAST_LOG_LEVEL_WARNING
-      fmt::print(stderr,
-                 "The option '{}' is deprecated for the '{}' subcommand and "
-                 "should be specified at the {} subcommand instead\n",
-                 opt, format, cmd);
-#endif
-      put(inv.options, "vast."s + cmd + '.' + opt, *x);
-    }
-  };
-  for (const auto& cmd : {"import", "spawn.source"})
-    for (const auto& format :
-         {"csv", "json", "suricata", "syslog", "test", "zeek", "zeek-json"})
-      for (const auto& opt :
-           {"listen", "read", "schema", "schema-file", "type", "uds"})
-        move_option_arg(cmd, format, opt);
-  for (const auto& cmd : {"export", "spawn.sink"})
-    for (const auto& format :
-         {"arrow", "ascii", "csv", "json", "null", "pcap", "zeek"})
-      for (const auto& opt : {"write", "uds"})
-        move_option_arg(cmd, format, opt);
-}
-
 caf::expected<invocation>
 parse(const command& root, command::argument_iterator first,
       command::argument_iterator last) {
@@ -402,7 +373,6 @@ parse(const command& root, command::argument_iterator first,
     manfooter(std::cout);
     return caf::no_error;
   }
-  fixup_options(result);
   return result;
 }